clean up, simplify

+++ abyss/dataset/data_restruct.py
@@ -44,15 +44,16 @@ class DataRestruct:
 
     def copy_helper(self, src, folder_name, case_name, tag_name):
         """Copy and renames files by their case and tag name, keeps file extension, returns the new file path"""
-        # if isinstance(src, str) and os.path.isfile(src):
-        file_name = os.path.basename(src)
-        file_extension = file_name.split(os.extsep, 1)[1]
-        new_file_name = f'{case_name}_{tag_name}.{file_extension}'
-        dst_file_path = os.path.join(
-            self.config_manager.params['project']['structured_dataset_store_path'],
-            folder_name,
-            new_file_name,
-        )
-        os.makedirs(os.path.dirname(dst_file_path), exist_ok=True)
-        shutil.copy2(src=src, dst=dst_file_path)
-        return dst_file_path
+        if isinstance(src, str) and os.path.isfile(src):
+            file_name = os.path.basename(src)
+            file_extension = file_name.split(os.extsep, 1)[1]  # split on first . and uses the rest as extension
+            new_file_name = f'{case_name}_{tag_name}.{file_extension}'
+            dst_file_path = os.path.join(
+                self.config_manager.params['project']['structured_dataset_store_path'],
+                folder_name,
+                new_file_name,
+            )
+            os.makedirs(os.path.dirname(dst_file_path), exist_ok=True)
+            shutil.copy2(src=src, dst=dst_file_path)
+            return dst_file_path
+        raise AssertionError(f'Files seems not to exist, case: {case_name}, tag: {tag_name}')
